==> 01-intro.Rmd <==
# Introduction {#intro}

This project is an exploration of the relationship between sentiment and the evaluations. Reddit, a social-engagement platform designed as a democratic forum of ideas, wherein each user can submit content of their own and evaluate that of other users. These forums are located within subreddits; subsections of the website dissociated according some symbolic purpose.  *Crypto-currencies* are an interesting area to study within this context because the block-chain technologies, upon which these "digital assets" are built, are democratic by there very nature. To speak precisely, the value of a crypto-currency "coin" is ultimately determined by the interdependent evaluations of its holders. In one of the early transactions made using the currency, 10000 [BTC](https://www.coindesk.com/price/bitcoin) was worth two pizzas; today, a single Bitcoin is worth a used car. This is surprising considering that, until very recently, these crypto-currencies were widely dismissed by global financial institutions; but as time goes on, [more governments, banks, and brokers](https://fortune.com/2020/07/22/bitcoin-big-banks-can-hold-occ-letter/) are finding value in this "intangible" asset.

Despite the hesitancy of major financial institutions in the early years of the technology, Bitcoin and other block-chain based currencies, often referred to as *Alt-Coins*, achieved confounding success in the later half of the 2010s. Since then it has fallen significantly, but today the price of Bitcoin is rapidly increasing and is on the verge of breaking it's previous record set in 2017. Regarding all of this, it is once again an appropriate time to ask: Why?

This project does not seek to answer this question, nor will it make make any empirical claims about the reliability or feasibility of using social media data to predict the value of currencies. Instead, this project is an exploratory analysis of data to assess the viability of using sentiment to predict the value of block-chain based currencies.

```{r, message = FALSE}
library(tidyverse)
library(tidytext)
library(textclean)
library(dplyr)
library(stringr)
library(knitr)
library(wordcloud)
library(kableExtra)
library(DT)
library(tidygraph)
library(ggraph)
library(tm)
library(ggwordcloud)
library(Hmisc)
library(lubridate)
library(wordcloud)
library(viridis)
library(colorspace)
library(Quandl)
library(reshape)
library(hrbrthemes)
library(scales)
library(lubridate)
library(plotly)

knitr::opts_chunk$set(echo = TRUE)
```
