==> 04-SentimentAnalysis.Rmd <==
# Sentiment Analysis {#sentiments}

```{r}
dataDir <- "Proj2Data"
```


```{r}
curTknsByMonth <- read.csv(paste0(dataDir, "/CoinTknsMonthly.csv"))
CommTkns <- read.csv(paste0(dataDir, "/CommTkns.csv"))
```


## Data Shaping

To group comments and their sentiments by `Coin` we have to first assign this identifier to the tokens via their associated `comm_id`.

```{r}
currTkns <- CommTkns %>%
  inner_join(curTknsByMonth[,c("Coin","Month","comm_id")], by = "comm_id")
```

### `get_sentiments()`

```{r}
nrc<- get_sentiments("nrc")
```

`r kbl(head(nrc))`

### Sentiment Counts By Coin

```{r}
#ggplot(sntmntByMonth, aes(x = Month, y = n, color = sentiment)) +
  #geom_line() +
  #facet_wrap(~Coin, scale = "free_y")

sBmPlot <- ggplot(sntmntByMonth, aes(x = as.Date(Month), y = n, group = sentiment, color = sentiment)) +
  geom_line() +
  scale_x_date(labels = date_format("%Y")) +
  facet_wrap(~Coin) +
  xlab("Month") + ylab("Number of Tokens (n)") 

ggplotly(sBmPlot)
```

```{r}
# Quandl limits free users to 50 calls a day, so these code chunks can be run for more recent data

#pBTC <- Quandl('BITFINEX/BTCUSD')
#pETH <- Quandl('BITFINEX/ETHUSD')
#pLTC <- Quandl('BITFINEX/LTCUSD')
#pXRP <- Quandl('BITFINEX/XRPUSD')

#pBTC["Coin"] <- "BTC"
#pETH["Coin"] <- "ETH"
#pLTC["Coin"] <- "LTC"
#pXRP["Coin"] <- "XRP"

#PricesByCoin <- rbind(pBTC, pETH, pLTC, pXRP)

#To get around this call limit, we use a csv to hold the data
PricesByCoin<- read.csv(paste0(dataDir, "/pricesByCoin.csv"))
```

```{r}


#write.csv(PricesByCoin,paste0(dataDir,"/pricesByCoin.csv"))
```

## Price & Sentiment over Time {.tabset}

```{r}

# function for getting Coin Price/ Sentiment vs Time graph by coin 

get_graph <- function(coin, coeff) {   # coin = "COIN_NAME", coeff = Value used to transform sentiment to match price scale on graph
  
  # get related coin data
  coinprice_data <- PricesByCoin %>% filter(Coin == coin)
  coin_sntmntByMonth <- sntmntByMonth %>% filter(Coin == coin)
  
  # reshape coin price by day data to merge high, low, last, med into one variable
  price_by_mkt_metric <- melt(coinprice_data, id = c("Date", "Coin"))
  colnames(price_by_mkt_metric)[3] <- "Mkt_Metrics"
  
  # normalize x-values for both datasets (date)
  price_by_mkt_metric$Date <- as_date(price_by_mkt_metric$Date)
  coin_sntmntByMonth$Month <- as_date(coin_sntmntByMonth$Month)
  

  # make the gg plot
  Coin_Daily_Price.plot <- price_by_mkt_metric %>%
    filter((Mkt_Metrics %in% c("High", "Low", "Last"))) %>%
    
    # ggplot setup
    ggplot(aes(x = Date)) +
    theme_minimal() +
    ggtitle(paste(coin, " Sentiment/ Price vs Time")) +
    xlab("Date") +
    theme(legend.title = element_blank())  +
    
    
    # plot price vs time lines
    geom_line(
      stat = 'identity',
      aes(
        y = value,
        linetype = Mkt_Metrics,
        color = Mkt_Metrics,
        size = Mkt_Metrics,
        alpha = Mkt_Metrics)) +
    scale_linetype_manual("Market Metrics", values = c("solid", "solid", "solid")) +
    scale_color_manual("Market Metrics", values = c('#EF9A9A', '#C5E1A5', '#212121')) +
    scale_size_manual("Market Metrics", values = c(1, 1, 0.3)) +
    scale_alpha_manual("Market Metrics", values = c(0.8, 0.8, 1)) +
    
    
    # plot sentiment bars (stacked)
    geom_bar(
      data = coin_sntmntByMonth,
      stat = 'identity',
      aes(
        x = Month,
        y = n / coeff,
        fill = sentiment)) +
    
    
    # setup y-axises
    scale_y_continuous(name = "Price (USD)",
                       sec.axis = sec_axis( ~ . * coeff,  name = "Sentiment (n)"))
  
  # convert to plotly
  Coin_Daily_Price.plotly = ggplotly(Coin_Daily_Price.plot, tooltip = c("label","x","y"))
  
  # cleans up ledgend labels
  for (i in 1:length(Coin_Daily_Price.plotly$x$data)) {
    if (!is.null(Coin_Daily_Price.plotly$x$data[[i]]$name)) {
      Coin_Daily_Price.plotly$x$data[[i]]$name =  gsub("\\(", "",
          str_split(Coin_Daily_Price.plotly$x$data[[i]]$name, ",")[[1]][1])
    }
  }
  
  Coin_Daily_Price.plot
  Coin_Daily_Price.plotly # FOR THE LIFE OF ME CANNOT FIGURE OUT HOW TO GET THE 2ND AXIS TO SHOW
}
```


### BTC

```{r}
get_graph("BTC", 4)
```

### ETH

```{r}
get_graph("ETH",20)
```

### LTC

```{r}
get_graph("LTC", 19)
```

### XRP

```{r}
get_graph("XRP", 8000)
```

## User Sentiment {.tabset}

```{r}
posts_by_coin <- currSntmntTkns %>% 
    group_by(Coin) %>%
    count(Coin)

users_by_coin <- currSntmntTkns %>% 
    group_by(user) %>%
    count(user)
```

```{r}
emo_stats <- function(emote, color1, color2){
  
  users.emo <- currSntmntTkns %>% 
    filter(sentiment == emote) %>%
    count(user) 
  users.emo <- inner_join(users.emo, users_by_coin, by = "user", suffix = c(".emote", ".total")) %>%
    filter(n.total >= 500) %>%
    mutate(n.emote_porp = n.emote/n.total) %>% 
    arrange(desc(n.emote_porp) )
  
  coin.emo <- currSntmntTkns %>% 
      filter(sentiment == emote) %>%
      group_by(Coin) %>% 
      count(Coin) 
  coin.emo <- inner_join(coin.emo, posts_by_coin, by = "Coin", suffix = c(".emote", ".total")) %>%
    mutate(n.emote_porp=n.emote/n.total)
  
  #users_coin.emo TOTAL plot
  plt1 <- ggplotly(
    ggplot(coin.emo, aes(x = Coin)) +
      theme_minimal() +
      theme(panel.grid.major.x = element_blank()) +
      geom_bar(aes(y = n.emote), stat='identity', fill = color1, width = 0.5) +
      ggtitle("Total emote by coin") +
      xlab("Coin") + ylab("Emote tokens (posts.emote)") 
  )
  
  # users_coin.emo PROP plot
  plt2 <- ggplotly(
    ggplot(coin.emo, aes(x = Coin)) +
      theme_minimal() +
      theme(panel.grid.major.x = element_blank()) +
      geom_bar(aes(y = n.emote_porp), stat='identity', fill = color2, width = 0.5) +
      ggtitle("Proportion of emote by coin") +
      xlab("Coin") + ylab("Emote tokens (posts.emote/ posts.total)") 
  )
  
  return(list(as_tibble(users.emo), plt1, plt2))
}
```



### Angriest
Angriest users
```{r}
ang <- emo_stats("anger", "#F1948A", "#E6B0AA")
head(ang[[1]])
ang[[2]]
ang[[3]]
```

### Happiest
Top most joyful/ positive users
```{r}
hap <- emo_stats(c("joy", "positive"), "#F7DC6F", "#FCF3CF")
head(hap[[1]])
hap[[2]]
hap[[3]]
```


### Saddest
Top saddest/ most negative users


```{r}
sad <- emo_stats(c("negative", "sad"), "#5499C7", "#D4E6F1")
head(sad[[1]])
sad[[2]]
sad[[3]]
```

## Descriptive Statistics

```{r}

ggplot(sntmntByMonth, aes(x = Coin, fill = Coin)) + geom_bar() +  ggtitle("Distribution of Coin Types")

ggplot(sntmntByMonth, aes(x = sentiment, fill = sentiment)) + geom_bar() +  ggtitle("Distribution Sentiments Among All Coins")


ggplot(sntmntByMonth, aes(x = Month, fill = Month)) + geom_bar() +  ggtitle("Number of Sentiment Types for Each Month")

ggplot(sntmntByMonth, aes(x = sentiment, y = n, color = Coin)) + geom_point() +  ggtitle("Popularity of Each Sentiment Among Coin Type")
```



## Most Abundant Sentiment Over Time

```{r}
aggSent <- function(pop.df, n, m, coindf, unevenStep = FALSE){
  k = 1
  j = 1
  for (i in 1:n){
    if(coindf[i,2] == m[j]){
      if(coindf[i,4] > pop.df[j,3]){
        pop.df[j,3] = coindf[i,4]
        pop.df[j,2] = coindf[i,3]
        pop.df[j,4] = pop.df[j,4] + coindf[i,4]
        pop.df[j,5] = pop.df[j,3] / pop.df[j,4]
      }
    }

    k = k + 1
    
    
    
    
    if(k == 11){
      j = j + 1
      k = 1
    }
  }
  return(pop.df)
}

df <- sntmntByMonth

df <- na.omit(df)

# create a new dataframe for each coin

btc <- df[which(df$Coin == "BTC"),]


btc_m <- unique(btc$Month)

btc_n <- nrow(btc)

n <- length(btc_m)
sentiment <- rep("x", n)


btc.pop.df <- data.frame(btc_m, sentiment,0, 0, 0)

btc.pop.df <- aggSent(btc.pop.df, btc_n, btc_m, btc)

# ETH

eth <- df[which(df$Coin == "ETH"),]
eth_m <-  unique(eth$Month)

eth_n <- nrow(eth)

n <- length(eth_m)
sentiment <- rep("x", n)

eth.pop.df <- data.frame(eth_m, sentiment,0, 0, 0)


#eth.pop.df <- aggSent(eth.pop.df, eth_n, eth_m, eth, unevenStep = TRUE)


#eth.pop.df

j <- 1

for(i in 1:eth_n){
  if(i > 1){
    prevM <- eth[i-1, 2]
    month <- eth[i,2]
  
    if(month != prevM){
      j <- j + 1
    }
    
  }
  
  if(eth[i,2] == eth_m[j]){
      if(eth[i,4] > eth.pop.df[j,3]){
        eth.pop.df[j,3] = eth[i,4]
        eth.pop.df[j,2] = eth[i,3]
        eth.pop.df[j,4] = eth.pop.df[j,4] + eth[i,4]
        eth.pop.df[j,5] = eth.pop.df[j,3] / eth.pop.df[j,4]
      }
    
  }
  
}



#xrp <- df[which(df$Coin == "XRP"),]

#eth.pop.df



ltc <- df[which(df$Coin == "LTC"),]

ltc_m <-  unique(ltc$Month)

ltc_n <- nrow(ltc)

n <- length(ltc_m)
sentiment <- rep("x", n)

ltc.pop.df <- data.frame(ltc_m, sentiment,0, 0, 0)

j <- 1

for(i in 1:ltc_n){
  if(i > 1){
    prevM <- ltc[i-1, 2]
    month <- ltc[i,2]
  
    if(month != prevM){
      j <- j + 1
    }
    
  }
  
  if(ltc[i,2] == ltc_m[j]){
      if(ltc[i,4] > ltc.pop.df[j,3]){
        ltc.pop.df[j,3] = ltc[i,4]
        ltc.pop.df[j,2] = ltc[i,3]
        ltc.pop.df[j,4] = ltc.pop.df[j,4] + ltc[i,4]
        ltc.pop.df[j,5] = ltc.pop.df[j,3] / ltc.pop.df[j,4]
      }
    
  }
  
  
}
ltc.pop.df

xrp <- df[which(df$Coin == "XRP"),]

xrp_m <-  unique(xrp$Month)

xrp_n <- nrow(xrp)

n <- length(xrp_m)
sentiment <- rep("x", n)

xrp.pop.df <- data.frame(xrp_m, sentiment,0, 0, 0)

j <- 1


for(i in 1:xrp_n){
  if(i > 1){
    prevM <- xrp[i-1, 2]
    month <- xrp[i,2]
  
    if(month != prevM){
      j <- j + 1
    }
  }
  if(xrp[i,2] == xrp_m[j]){
      if(xrp[i,4] > xrp.pop.df[j,3]){
        xrp.pop.df[j,3] = xrp[i,4]
        xrp.pop.df[j,2] = xrp[i,3]
        xrp.pop.df[j,4] = xrp.pop.df[j,4] + xrp[i,4]
        xrp.pop.df[j,5] = xrp.pop.df[j,3] / xrp.pop.df[j,4]
      }
  }
}
xrp.pop.df
```

## Plots of Top Sentiment Over Time

```{r}
ggplot(data = btc.pop.df, aes(x=btc_m, y = X0.2, group = 1, color = sentiment))+
  geom_line()+
  geom_point() + 
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) + ggtitle("Most Common Sentiment Over Time For BTC") + xlab("Month") + ylab("Proportion of Sentiment")

ggplot(data = eth.pop.df, aes(x=eth_m, y = X0.2, group = 1, color = sentiment))+
  geom_line()+
  geom_point() + 
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) + ggtitle("Most Common Sentiment Over Time For ETH") + xlab("Month") + ylab("Proportion of Sentiment")

ggplot(data = ltc.pop.df, aes(x=ltc_m, y = X0.2, group = 1, color = sentiment))+
  geom_line()+
  geom_point() + 
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) + ggtitle("Most Common Sentiment Over Time For LTC") + xlab("Month") + ylab("Proportion of Sentiment")


ggplot(data = xrp.pop.df, aes(x=xrp_m, y = X0.2, group = 1, color = sentiment))+
  geom_line()+
  geom_point() + 
  theme(axis.text.x = element_text(angle = 60, hjust = 1)) + ggtitle("Most Common Sentiment Over Time For XRP") + xlab("Month") + ylab("Proportion of Sentiment")
```
