==> 05-references.Rmd <==
`r if (knitr:::is_html_output()) '
# References {-}

'`
## Packages Used

```{r include=TRUE}
# automatically create a bib database for R packages
knitr::write_bib(c(
  .packages(), 'bookdown', 'knitr', 'rmarkdown','packages.bib', "book.bib"))
```
